+++ test/api/topic.js
@@ -441,68 +441,37 @@ var topicCommentEdit = function (agent, userId, topicId, commentId, subject, tex
     _topicCommentEdit(agent, userId, topicId, commentId, subject, text, type, 200, callback);
 };
 
-var _topicCommentList = function (agent, userId, topicId, expectedHttpCode, callback) {
+var _topicCommentList = function (agent, userId, topicId, orderBy, expectedHttpCode, callback) {
     var path = '/api/users/:userId/topics/:topicId/comments'
         .replace(':userId', userId)
         .replace(':topicId', topicId);
 
     agent
         .get(path)
-        .expect(expectedHttpCode)
-        .expect('Content-Type', /json/)
-        .end(callback);
-};
-
-var topicCommentList = function (agent, userId, topicId, callback) {
-    _topicCommentList(agent, userId, topicId, 200, callback);
-};
-
-var _topicCommentListV2 = function (agent, userId, topicId, orderBy, expectedHttpCode, callback) {
-    var path = '/api/v2/users/:userId/topics/:topicId/comments'
-        .replace(':userId', userId)
-        .replace(':topicId', topicId);
-
-    agent
-        .get(path)
         .query({orderBy: orderBy})
         .expect(expectedHttpCode)
         .expect('Content-Type', /json/)
         .end(callback);
 };
 
-var topicCommentListV2 = function (agent, userId, topicId, orderBy, callback) {
-    _topicCommentListV2(agent, userId, topicId, orderBy, 200, callback);
+var topicCommentList = function (agent, userId, topicId, orderBy, callback) {
+    _topicCommentList(agent, userId, topicId, orderBy, 200, callback);
 };
 
-var _topicCommentListUnauth = function (agent, topicId, expectedHttpCode, callback) {
+var _topicCommentListUnauth = function (agent, topicId, orderBy, expectedHttpCode, callback) {
     var path = '/api/topics/:topicId/comments'
         .replace(':topicId', topicId);
 
     agent
         .get(path)
         .expect(expectedHttpCode)
-        .expect('Content-Type', /json/)
-        .end(callback);
-};
-
-var topicCommentListUnauth = function (agent, topicId, callback) {
-    _topicCommentListUnauth(agent, topicId, 200, callback);
-};
-
-var _topicCommentListUnauthV2 = function (agent, topicId, orderBy, expectedHttpCode, callback) {
-    var path = '/api/v2/topics/:topicId/comments'
-        .replace(':topicId', topicId);
-
-    agent
-        .get(path)
-        .expect(expectedHttpCode)
         .query({orderBy: orderBy})
         .expect('Content-Type', /json/)
         .end(callback);
 };
 
-var topicCommentListUnauthV2 = function (agent, topicId, orderBy, callback) {
-    _topicCommentListUnauthV2(agent, topicId, orderBy, 200, callback);
+var topicCommentListUnauth = function (agent, topicId, orderBy, callback) {
+    _topicCommentListUnauth(agent, topicId, orderBy, 200, callback);
 };
 
 var _topicCommentDelete = function (agent, userId, topicId, commentId, expectedHttpCode, callback) {
@@ -6654,7 +6623,6 @@ suite('Users', function () {
                         assert.equal(comment.creator.id, user.id);
 
                         var commentReplyText = 'Child comment';
-
                         topicCommentCreate(agent, user.id, topic.id, comment.id, comment.edits.length - 1, null, null, commentReplyText, function (err, res) {
                             if (err) return done(err);
 
@@ -6666,7 +6634,6 @@ suite('Users', function () {
                             assert.equal(commentReply.text, commentReplyText);
                             assert.equal(commentReply.creator.id, user.id);
                             assert.equal(commentReply.parent.id, comment.id);
-
                             done();
                         });
                     });
@@ -6751,7 +6718,7 @@ suite('Users', function () {
 
                             var status = res.body.status;
                             assert.equal(status.code, 20000);
-                            topicCommentListV2(agent3, user3.id, topic.id, 'date', function (err, res) {
+                            topicCommentList(agent3, user3.id, topic.id, 'date', function (err, res) {
                                 if (err) return done(err);
 
                                 var commentEdited = res.body.data.rows[0];
@@ -6837,7 +6804,7 @@ suite('Users', function () {
                 });
 
                 test('Success', function (done) {
-                    topicCommentList(agent, user.id, topic.id, function (err, res) {
+                    topicCommentList(agent, user.id, topic.id, null, function (err, res) {
                         if (err) return done(err);
 
                         var list = res.body.data;
@@ -6848,7 +6815,7 @@ suite('Users', function () {
                         delete creatorExpected.imageUrl; // Image url is not returned as it's not needed for now
                         delete creatorExpected.language; // Language is not returned
 
-                        assert.equal(list.count, 2);
+                        assert.equal(list.count.total, 2);
                         assert.equal(comments.length, 2);
 
                         // Comment 1
@@ -6880,7 +6847,7 @@ suite('Users', function () {
                 });
 
                 test('Success v2', function (done) {
-                    topicCommentListV2(agent, user.id, topic.id, 'rating', function (err, res) {
+                    topicCommentList(agent, user.id, topic.id, 'rating', function (err, res) {
                         if (err) return done(err);
 
                         var list = res.body.data;
@@ -6891,7 +6858,7 @@ suite('Users', function () {
                         delete creatorExpected.imageUrl; // Image url is not returned as it's not needed for now
                         delete creatorExpected.language; // Language is not returned
 
-                        assert.equal(list.count, 2);
+                        assert.equal(list.count.total, 2);
                         assert.equal(comments.length, 2);
 
                         // Comment 1
@@ -6931,7 +6898,7 @@ suite('Users', function () {
                             function (err) {
                                 if (err) return done(err);
 
-                                topicCommentList(agent, user.id, topic.id, function (err, res) {
+                                topicCommentList(agent, user.id, topic.id, null, function (err, res) {
                                     if (err) return done(err);
 
                                     var list = res.body.data;
@@ -6942,7 +6909,7 @@ suite('Users', function () {
                                     delete creatorExpected.imageUrl; // Image url is not returned, as it's not needed for now
                                     delete creatorExpected.language; // Language is not returned
 
-                                    assert.equal(list.count, 2);
+                                    assert.equal(list.count.total, 2);
                                     assert.equal(comments.length, 2);
 
                                     // Comment 1
@@ -7033,7 +7000,7 @@ suite('Users', function () {
                                 });
                         })
                         .then(function () {
-                            topicCommentList(agent, user.id, topic.id, function (err, res) {
+                            topicCommentList(agent, user.id, topic.id, null, function (err, res) {
                                 if (err) return done(err);
 
                                 var replyText11 = 'R1.1';
@@ -7049,7 +7016,7 @@ suite('Users', function () {
                                 delete creatorExpected.imageUrl; // Image url is not returned, as it's not needed for now
                                 delete creatorExpected.language; // Language is not returned
 
-                                assert.equal(list.count, 2);
+                                assert.equal(list.count.total, 2);
                                 assert.equal(comments.length, 2);
 
                                 // Comment 1
@@ -7154,12 +7121,13 @@ suite('Users', function () {
                     topicCommentDelete(agent, user.id, topic.id, comment.id, function (err) {
                         if (err) return done(err);
 
-                        topicCommentList(agent, user.id, topic.id, function (err, res) {
+                        topicCommentList(agent, user.id, topic.id, null, function (err, res) {
                             if (err) return done(err);
 
                             var comments = res.body.data;
-                            assert.equal(comments.count, 0);
-                            assert.equal(comments.rows.length, 0);
+                            assert.equal(comments.count.total, 1);
+                            assert.equal(comments.rows.length, 1);
+                            assert.isNotNull(comments.rows[0].deletedAt);
 
                             done();
                         });
@@ -8655,10 +8623,10 @@ suite('Topics', function () {
                             .parallel(
                                 [
                                     function (cb) {
-                                        topicCommentList(creatorAgent, creator.id, topic.id, cb);
+                                        topicCommentList(creatorAgent, creator.id, topic.id, null, cb);
                                     },
                                     function (cb) {
-                                        topicCommentListUnauth(userAgent, topic.id, cb);
+                                        topicCommentListUnauth(userAgent, topic.id, null, cb);
                                     }
                                 ],
                                 function (err, results) {
@@ -8687,10 +8655,10 @@ suite('Topics', function () {
                         .parallel(
                             [
                                 function (cb) {
-                                    topicCommentList(creatorAgent, creator.id, topic.id, cb);
+                                    topicCommentList(creatorAgent, creator.id, topic.id, null, cb);
                                 },
                                 function (cb) {
-                                    topicCommentListUnauth(userAgent, topic.id, cb);
+                                    topicCommentListUnauth(userAgent, topic.id, null, cb);
                                 }
                             ],
                             function (err, results) {
@@ -8756,12 +8724,16 @@ suite('Topics', function () {
                     count: 3,
                     rows: [reply1, reply2, reply3]
                 };
-                topicCommentListUnauthV2(userAgent, topic.id, 'date', function (err, res) {
+                topicCommentListUnauth(userAgent, topic.id, 'date', function (err, res) {
                     if (err) return done(err);
                     var data = res.body.data;
                     var expectedResult = {
                         rows: [comment3, comment2, comment1],
-                        count: 3
+                        count: {
+                            total: 3,
+                            pro: 2,
+                            con: 1
+                        }
                     };
                     assert.shallowDeepEqual(data, expectedResult);
                     done();
@@ -8818,12 +8790,16 @@ suite('Topics', function () {
                     count: 3,
                     rows: [reply3, reply1, reply2]
                 };
-                topicCommentListV2(creatorAgent, creator.id, topic.id, 'rating', function (err, res) {
+                topicCommentList(creatorAgent, creator.id, topic.id, 'rating', function (err, res) {
                     if (err) return done(err);
                     var data = res.body.data;
                     var expectedResult = {
                         rows: [comment2, comment3, comment1],
-                        count: 3
+                        count: {
+                            total: 3,
+                            pro: 2,
+                            con: 1
+                        }
                     };
                     assert.shallowDeepEqual(data, expectedResult);
                     done();
@@ -8880,12 +8856,16 @@ suite('Topics', function () {
                     count: 3,
                     rows: [reply2, reply3, reply1]
                 };
-                topicCommentListV2(creatorAgent, creator.id, topic.id, 'popularity', function (err, res) {
+                topicCommentList(creatorAgent, creator.id, topic.id, 'popularity', function (err, res) {
                     if (err) return done(err);
                     var data = res.body.data;
                     var expectedResult = {
                         rows: [comment2, comment1, comment3],
-                        count: 3
+                        count: {
+                            total: 3,
+                            pro: 2,
+                            con: 1
+                        }
                     };
                     assert.shallowDeepEqual(data, expectedResult);
                     done();
@@ -8961,13 +8941,17 @@ suite('Topics', function () {
                             });
                     })
                     .then(function () {
-                        topicCommentListV2(creatorAgent, creator.id, topic.id, 'date', function (err, res) {
+                        topicCommentList(creatorAgent, creator.id, topic.id, 'date', function (err, res) {
                             if (err) return done(err);
 
                             var data = res.body.data;
                             var expectedResult = {
                                 rows: [comment3, comment2, comment1],
-                                count: 3
+                                count: {
+                                    total: 3,
+                                    pro: 2,
+                                    con: 1
+                                }
                             };
 
                             assert.shallowDeepEqual(data, expectedResult);
@@ -8992,7 +8976,7 @@ suite('Topics', function () {
 
                     var topic = res.body.data;
 
-                    _topicCommentListUnauth(userAgent, topic.id, 404, function (err) {
+                    _topicCommentListUnauth(userAgent, topic.id, null, 404, function (err) {
                         if (err) return done(err);
 
                         done();